Fix macos copy & paste
To allow copy and paste to work in macos, I needed to install
a utility program reattach-to-user-namespace. I then needed
to configure tmux to use this program when creating a new
session and when calling the pbcopy and pbpaste commands.
I created a seperate tmux config file for darwin specific
configuration and only configure the bindings when the utility
is installed.