Fix hang when using tmux 2.4+
The format of the bind-key command has changed and this was causing tmux
to hang at startup in iTerm2 when tmux was started with the -CC option.
I added a check to the config file for the tmux version and am now using
the new bind-key format on tmux 2.4 and greater.